refactor pipeline. resolve task <-> sandbox relationship
# Conflicts:
#	src/language_model/data/extract.py
#	src/language_model/data/load.py
#	src/language_model/modelling/trainer.py
#	src/language_model/tokenization/extend.py
#	src/language_model/tokenization/tasks.py
#	src/language_model/tokenization/trainer.py
